fix(error-feed): render the Overview fix as markdown, not as its own source
The cluster analysis card printed `data.fix` through a bare Typography, so the
model's markdown arrived on screen verbatim — backticks, asterisks and the diff's
`+` markers run together in one paragraph. The Fix tab renders the same string
properly, so one fix looked like two different things depending on the tab.

Move AnalyzeMarkdown out of AnalyzeTab into its own module and use it in both
places, rather than growing a second renderer that would drift from the first.

No container of its own: the analysis card around it already is one, and the Fix
tab lays it out the same way — only the diff gets a box, from the renderer's
`pre` styling. minWidth: 0 on the markdown column so a long unbroken diff line
scrolls inside that box instead of widening the whole card.
